remove include where is not necessary

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rzamolo- <rzamolo-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/13 15:09:32 by rzamolo-          #+#    #+#             */
-/*   Updated: 2022/09/20 12:55:52 by rzamolo-         ###   ########.fr       */
+/*   Updated: 2022/09/22 18:05:49 by rzamolo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned long long	length;
+	size_t	length;
 
 	length = 0;
 	while (*s != '\0')
